Get rid of ValueM and add Monoid instances
The issue that made it necessary to introduce ValueM (a wrapper around
`Aeson.Value` with a monoid structure) was that the `Alternative` instance for
`SchemaOut` (hence `SchemaP`) requires a monoid constraint on the output value.
The constraint is not really necessary for the definition of the instance, but
it is required because `Alternative` is a subclass of `Applicative`, and the
instance for the latter needs it.

I have now fixed this little annoyance by deriving `Monoid` instances for all
the profunctors involved, including `SchemaP`. Now one can simply use `<>`
instead of `<|>` when defining alternatives for schemas. They have the exact
same behaviour, so nothing is actually changed, but now the output value does
not need to be a monoid, which means we can just use `Aeson.Value` directly.